fix: move clerkThrottle into src/ so it compiles to dist/
#1530 imported clerkThrottle from seed/util/, but seed/ lives outside
nest-cli's sourceRoot. SWC only compiles src/, so dist/seed/ never gets
emitted — node dist/main.js crashes on UsersService load with
"Cannot find module ../../../seed/util/clerkThrottle.util".

The throttle utilities aren't seed-specific anyway; move them to
src/vendors/clerk/util/ where the rest of the Clerk integration lives,
and update the three importers (seed/users.ts and scripts/migrate-clean.ts
keep working under tsx, which resolves the new path the same way).
